+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## [0.0.3] - SNAPSHOT
 
+### New
+- Add `brightness` command for setting brightness.
+E.g.: `wled_cli brightness localhost 255` - set max brightness.
+
+- Add `inc_brightness` command for increasing brightness.
+E.g.: `wled_cli inc_brightness localhost 10` - increase brightness by 10 points (step value can be omitted).
+
+- Add `dec_brightness` command for decreasing brightness.
+E.g.: `wled_cli dec_brightness localhost 10` - deccrease brightness by 10 points (step value can be omitted).
+
 
 ## [0.0.2] - 29 Jan 2024